+++ mobius/Cargo.toml
@@ -8,6 +8,7 @@ abstraction = { path = "../abstraction" }
 svg = "0.18.0"
 
 [dev-dependencies]
+clap = { version = "4.5.20", features = ["derive"] }
 test-case = "3.3.1"
 pretty_assertions = "1.4.1"
