--- mobius/examples/ghost_tour.rs
@@ -1,127 +1,0 @@
-use std::f64::consts::{FRAC_PI_2, FRAC_PI_3, PI};
-
-use abstraction::Group;
-use mobius::{
-    algorithms::{GridIFS, GroupIFS, SemigroupIFS},
-    elliptic, loxodromic, map_triple,
-    motifs::ghost,
-    rotation, scale,
-    svg_plot::{render_views, style_geometry, View},
-    transformable::Transformable,
-    translation, Complex, Mobius,
-};
-
-pub fn main() -> Result<(), std::io::Error> {
-    let (ghost, ghost_style) = ghost();
-
-    // Show the ghost by themself -----------------------------------
-    render_views(
-        "output",
-        "ghosty",
-        &[View("", 0.0, -0.5, 2.5)],
-        style_geometry(ghost_style, &ghost),
-    )?;
-
-    // Oh no! the ghost fell down the drain! ----------------------
-
-    // Create a transform with a spiral sink at +3
-    let translate3 = translation(3.0.into()).unwrap();
-    let spiral_in = rotation(-FRAC_PI_3).unwrap() * scale(0.6).unwrap();
-    let drain = Mobius::sandwich(translate3, spiral_in);
-    let ifs = SemigroupIFS::new(vec![drain]);
-    let down_the_drain = ifs.apply(&ghost, 0, 20);
-
-    render_views(
-        "output",
-        "ghost_down_drain",
-        &[View("", 1.5, 0.0, 2.5)],
-        style_geometry(ghost_style, &down_the_drain[..]),
-    )?;
-
-    // Caught between parabolic transforms that have fixed points at -1 and
-    // 1, and map the opposite point on the unit circle 90 degrees around the
-    // circle. This would benefit from skipping inverses, parts of this
-    // diagram will be rendered many times.
-    // -------------------------------
-    let shrink = scale(0.125).unwrap();
-    let small_ghost = ghost.transform(shrink);
-    let left_parabolic = map_triple(
-        (-Complex::ONE, Complex::ONE, -Complex::I),
-        (-Complex::ONE, Complex::I, Complex::ONE),
-    )
-    .unwrap();
-    let right_parabolic = map_triple(
-        (Complex::ONE, -Complex::ONE, -Complex::I),
-        (Complex::ONE, Complex::I, -Complex::ONE),
-    )
-    .unwrap();
-
-    let parabolic_ifs = GroupIFS::new(vec![left_parabolic, right_parabolic]);
-    let parabolic_walk = parabolic_ifs.apply(&small_ghost, 0, 5);
-    render_views(
-        "output",
-        "ghost_parabolic",
-        &[View("", 0.0, 0.0, 0.9), View("zoom_in", -0.5, 0.1, 0.2)],
-        style_geometry(ghost_style, &parabolic_walk[..]),
-    )?;
-
-    let parabolic_grid = GridIFS::new(vec![(left_parabolic, -20, 20), (right_parabolic, -20, 20)]);
-    let parabolic_grid_walk = parabolic_grid.apply(&small_ghost);
-    render_views(
-        "output",
-        "ghost_parabolic_grid",
-        &[View("", 0.0, 0.0, 1.1)],
-        style_geometry(ghost_style, &parabolic_grid_walk[..]),
-    )?;
-
-    // A loxodromic double spiral. Though instead of going from -1 to 1,
-    // I want it from -i to i, so conjugate by a rotate
-    let double_spiral = loxodromic(Complex::new(1.5, 1.1)).unwrap();
-    let rotate90 = rotation(FRAC_PI_2).unwrap();
-    let vertical_spiral = Mobius::sandwich(rotate90, double_spiral);
-    let ellip4 = elliptic(FRAC_PI_2).unwrap();
-    let vertical_ellip = Mobius::sandwich(rotate90, ellip4);
-
-    let grid = GridIFS::new(vec![(vertical_spiral, -8, 8), (vertical_ellip, 0, 4)]);
-
-    //let ifs = SemigroupIFS::new(vec![vertical_spiral, vertical_spiral.inverse()]);
-    //let double_spiral_walk = ifs.apply(&small_ghost, 0, 10);
-    let double_spiral_walk = grid.apply(&small_ghost);
-
-    render_views(
-        "output",
-        "ghost_double_spiral",
-        &[View("", 0.0, 0.0, 1.0), View("sink", -0.125, 0.75, 0.5)],
-        style_geometry(ghost_style, &double_spiral_walk[..]),
-    )?;
-
-    // Two 90 degree elliptic rotations 90 degrees apart. This is isomorphic
-    // to the rotation symmetry group of the cube/octahedron
-    let swirl = elliptic(FRAC_PI_2).unwrap();
-    let swirl2 = Mobius::sandwich(rotate90, swirl);
-    let to_the_left = translation(Complex::new(-0.5, 0.0)).unwrap();
-    let shifted_ghost = small_ghost.transform(to_the_left);
-    let ifs = SemigroupIFS::new(vec![swirl, swirl2]);
-    let swirl_walk = ifs.apply(&shifted_ghost, 0, 8);
-    render_views(
-        "output",
-        "ghost_octahedral",
-        &[View("", 0.0, 0.0, 3.0)],
-        style_geometry(ghost_style, &swirl_walk[..]),
-    )?;
-
-    // But now if we make the rotation slightly different, things don't
-    // quite line up. I find the result amusing.
-    let swirl = elliptic(PI / 2.01).unwrap();
-    let swirl2 = Mobius::sandwich(rotate90, swirl);
-    let ifs = SemigroupIFS::new(vec![swirl, swirl2]);
-    let swirl_walk = ifs.apply(&shifted_ghost, 0, 8);
-    render_views(
-        "output",
-        "ghost_triggered",
-        &[View("", 0.0, 0.0, 3.0)],
-        style_geometry(ghost_style, &swirl_walk[..]),
-    )?;
-
-    Ok(())
-}
